requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a1676+g3f06f24
idna==2.7
idna-ssl==1.1.0
multidict==4.3.1
pipenv==2018.7.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.18
pycrypto==2.6.1
PyNaCl==1.1.2
requests==2.21.0
six==1.11.0
urllib3==1.24.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
websockets==6.0
yarl==1.2.6
